+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ## Version 4.1.0
 * Support for revoke of CLPv2 license for autoscaling solutions that license from BIG-IQ 5.3+
+* Support for licensing an unreachable device from BIG-IQ 5.4+
+    * --cloud parameter is required when calling onboard.js when licensing via BIG-IQ 5.4+
 * More reliable provisioning of ASM and AFM
 
 ## Version 4.0.0